backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
google-generativeai
numpy>=1.24.0
python-dotenv>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0
